pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
description="flask template"
[tool.poetry.dependencies]
flask = "==1.0.2"
python = "^3.8"
python-dotenv = "^0.15.0"
werkzeug = "^1.0.1"
flask-login = "^0.5.0"
flask-sqlalchemy = "^2.4.4"
flask-sslify = "^0.1.5"
